requirements.txt
black==20.8b1
Django==3.1.1
django-crispy-forms==1.9.2
djangorestframework==3.11.1
pytest-django==3.9.0
python-dotenv==0.14.0
tox==3.20.0
Pillow==7.2.0
drf-nested-routers==0.91
